Introduced the recommendation to use DEFERRED_EXPLICIT
The introduction means that unless the developer specifically wants
the entity manager to flush changes to CommentBundle entities, that
the flush will only occur when CommentBundle intends the flushes to
occur.

It is not necessary, but recommended due to the nature of the ORM
EntityManager flushing all changes when requested, rather than just
changes we want to be flushed.

Fixes #26